feat: horizontal variant
